Change BitmapText to use baseTexture (the font holding pbSurface) and a cellFrame number, instead of creating dozens or hundreds of tiny surfaces one per character.  This will hugely speed up the webgl performance as all bitmap text will go through the batch renderer.
Modified LoadTexture to call setFrame when the key is pbSurface, and permit a number for the 'frame' parameter as well as the previous Phaser.Frame structure.

Removed some dead code from LoaderParser, added the cellFrame into the data.chars object for each letter in the font, and called the new pbSurface method setCellFrame to establish all the variables required by the batch renderer to blit that single letter from the font texture.

pbSurface - added setCellFrame method, also fixed misconception about the purpose of cellOffsets (it's an offset to the render location of each animation frame, NOT the offset to the corner of the texture fragment that is to be rendered... that value is represented by cellTextureBounds as a fraction of the entire source texture).

This permits examples | audio | audio sprite duration to display bitmapfonts on the sound effect buttons.
TODO: the alignment is currently 'left' and it should be 'center'